+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ### Unreleased
 
+* Rename `GWTGzipFilter` to `PreEncodedGzipFilter` to reflect reality that filter is unrelated to GWT.
+* Introduce a `AbstractPreEncodedFilter` base class to make supporting other pre-encoded formats easy.
+* Add a `PreEncodedBrotliFilter` so support serving files pre-encoded with brotli.
+
 ### [v0.8](https://github.com/realityforge/gwt-cache-filter/tree/v0.8) (2020-03-06) · [Full Changelog](https://github.com/realityforge/gwt-cache-filter/compare/v0.7...v0.8)
 
 Changes in this release: